Completed assignment

+++ Binomial.java
@@ -11,18 +11,18 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, basic version.
-	public static int binomial1(int n, int k) { 
+	public static long binomial1(long n, long k) { 
 		if (k > n) return 0;
 		if (k == 0 || n == 0) return 1;
 		return binomial1(n - 1, k) + binomial1(n - 1, k - 1);
 	 }
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(long n, long k) {
 		//// This function creates a 2D array, say memo, 
 		//// and then initializes all its elements to -1.
 		//// It then calls binomial(n, k, memo), which does all the heavy lifiting.
-		int[][] memo = new int[n + 1][k + 1];
+		long[][] memo = new long[n + 1][ k + 1];
 		for (int i = 0; i <= n; i++) {
 			for (int j = 0; j <= k; j++) {
 				memo[i][j] = -1;
@@ -31,7 +31,7 @@ public class Binomial {
 		return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(long n, long k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
